i2c: s3c2410: stop all pending transfers on shutdown
If the i2c bus is busy when reboot() is called, it may never
come out of the busy state. Usually this is not an issue - since
the hardware will be reset before another kernel comes along - but
this causes the i2c bus to be unusable when using kexec.